backend/requirements.txt
fastapi
uvicorn
transformers
torch
pydantic 
firebase-admin
openai
python-dotenv